refactor(runtime): isolate package plan manifest emission
