+++ CMakeLists.txt
@@ -526,6 +526,7 @@ set(COPPERFIN_RUNTIME_PIPELINE_SOURCES
     src/runtime/runtime_pipeline_package_content_io.cpp
     src/runtime/runtime_pipeline_manifest_pair.cpp
     src/runtime/runtime_pipeline_manifest_pair_io.cpp
+    src/runtime/runtime_pipeline_package_plan_manifest.cpp
     src/runtime/runtime_pipeline_public_api.cpp
 )
